requirements.txt
astroid==4.0.2
autopep8==2.3.2
beets==2.5.1
certifi==2025.11.12
charset-normalizer==3.4.4
confuse==2.1.0
dill==0.4.0
filetype==1.2.0
idna==3.11
isort==7.0.0
jellyfish==1.2.1
lap==0.5.12
mccabe==0.7.0
mediafile==0.13.0
musicbrainzngs==0.7.1
mutagen==1.47.0
numpy==2.3.5
platformdirs==4.5.0
pycodestyle==2.14.0
pylint==4.0.3
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
slskd-api==0.2.1
tomlkit==0.13.3
typing_extensions==4.15.0
Unidecode==1.4.0
urllib3==2.5.0
